requirements_web.txt
flask>=3.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
